+++ std-bits/table/src/main/java/org/enso/table/data/table/join/JoinStrategy.java
@@ -1,10 +1,12 @@
 package org.enso.table.data.table.join;
 
+import org.enso.table.data.table.join.between.SortJoin;
 import org.enso.table.data.table.join.conditions.Between;
 import org.enso.table.data.table.join.conditions.Equals;
 import org.enso.table.data.table.join.conditions.EqualsIgnoreCase;
 import org.enso.table.data.table.join.conditions.HashableCondition;
 import org.enso.table.data.table.join.conditions.JoinCondition;
+import org.enso.table.data.table.join.hashing.HashJoin;
 import org.enso.table.problems.ProblemAggregator;
 
 import java.util.List;
@@ -37,9 +39,9 @@ public interface JoinStrategy {
       assert !betweenConditions.isEmpty();
       return new SortJoin(betweenConditions);
     } else if (betweenConditions.isEmpty()) {
-      return new EqualityHashJoin(hashableConditions, new MatchAllStrategy());
+      return new HashJoin(hashableConditions, new MatchAllStrategy());
     } else {
-      return new EqualityHashJoin(hashableConditions, new SortJoin(betweenConditions));
+      return new HashJoin(hashableConditions, new SortJoin(betweenConditions));
     }
   }
